# Optimizations for threadpool