… CHANGES.rst
… benchmarks/bench_pool.py