@jamadden 2018-03-23
No, in some pathological cases the wait order can make a performance difference, when max size is less than the iterate size.