@jamadden 2018-03-24
Final numbers for plain greenlet pools:

| Benchmark          | 36_pool_master | 36_pool_opts                |
|--------------------|----------------------|------------------------------|
| imap_unordered_seq | 803 us         | 686 us: 1.17x faster (-15%) |
| imap_unordered_par | 445 us         | 389 us: 1.14x faster (-13%) |
| imap_seq           | 793 us         | 729 us: 1.09x faster (-8%)  |
| imap_par           | 407 us         | 398 us: 1.02x faster (-2%)  |
| map_seq            | 715 us         | 293 us: 2.44x faster (-59%) |
| map_par            | 388 us         | 199 us: 1.96x faster (-49%) |

Not significant: apply, spawn

And the threadpool:

| Benchmark          | 36_threadpool_master | 36_threadpool_opt_cond10     |
|--------------------|----------------------|------------------------------|
| imap_unordered_seq | 1.15 ms              | 1.02 ms: 1.13x faster (-11%) |
| imap_unordered_par | 1.02 ms              | 928 us: 1.10x faster (-9%)   |
| imap_seq           | 1.17 ms              | 1.03 ms: 1.13x faster (-11%) |
| imap_par           | 1.07 ms              | 993 us: 1.08x faster (-7%)   |
| map_seq            | 1.16 ms              | 870 us: 1.33x faster (-25%)  |
| map_par            | 1.07 ms              | 675 us: 1.59x faster (-37%)  |
| apply              | 1.22 ms              | 1.12 ms: 1.10x faster (-9%)  |
| spawn              | 1.21 ms              | 1.09 ms: 1.11x faster (-10%) |